requirements.txt
future==0.18.2
numpy==1.18.5
Pillow==8.1.1
PyYAML==5.4
yacs==0.1.7
scikit-image
Cython
tqdm

